code cleanup

+++ app/models/image.rb
@@ -13,17 +13,12 @@ class Image < ActiveRecord::Base
   scope :of_kind, ->(kind){ where(kind: kind) }
 
   def self.update_ids(contest_id, image_ids, kind)
-    old_ids = of_kind(kind).pluck(:id)
-    ids_to_remove = old_ids - image_ids
-    ids_to_add = image_ids - old_ids
-    where(id: ids_to_remove).destroy_all
-    add_images(contest_id, ids_to_add, kind)
-  end
-
-  def self.add_images(contest_id, image_ids, kind)
-    where(id: image_ids).each do |image|
-      image.update_attributes(contest_id: contest_id, kind: kind)
+    transaction do
+      old_ids = of_kind(kind).pluck(:id)
+      ids_to_remove = old_ids - image_ids
+      where(id: ids_to_remove).destroy_all if ids_to_remove.present?
+      ids_to_add = image_ids - old_ids
+      where(id: ids_to_add).update_all(contest_id: contest_id, kind: kind) if ids_to_add.present?
     end
   end
-
 end
